Rename EventUtility to WebhookUtility
